pyproject.toml
[tool.poetry]
name = "animal-detection-hack"
version = "0.1.0"
description = ""
authors = ["Роман Лазовский <r.v.lazovskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
pyqt5 = "^5.15.10"
pillow = "^10.3.0"
ultralytics = "^8.2.16"
matplotlib = "^3.9.0"
certifi = "^2024.2.2"
charset-normalizer = "^3.3.2"
contourpy = "^1.2.1"
cycler = "^0.12.1"
filelock = "^3.14.0"
fonttools = "^4.51.0"
fsspec = "^2024.5.0"
idna = "^3.7"
jinja2 = "^3.1.4"
kiwisolver = "^1.4.5"
markupsafe = "^2.1.5"
mpmath = "^1.3.0"
networkx = "^3.3"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
packaging = "^24.0"
pandas = "^2.2.2"
psutil = "^5.9.8"
py-cpuinfo = "^9.0.0"
pyparsing = "^3.1.2"
pyqt5-qt5 = "^5.15.13"
pyqt5-sip = "^12.13.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
scipy = "^1.13.0"
seaborn = "^0.13.2"
six = "^1.16.0"
sympy = "^1.12"
thop = "^0.1.1.post2209072238"
torch = "^2.3.0"
torchvision = "^0.18.0"
tqdm = "^4.66.4"
typing-extensions = "^4.11.0"
tzdata = "^2024.1"
urllib3 = "^2.2.1"
pyinstaller = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
